requirements.txt
# 
fastapi[standard]==0.125.0
uvicorn[standard]==0.30.0

#
python-dotenv==1.2.0
pydantic-settings==2.12.0

# Environment and configuration
python-dotenv==1.2.0
pydantic-settings>=2.8.0

# HTTP and form handling
python-multipart==0.0.20
httpx==0.28.1

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0


mindsdb_sdk
supabase>=2.13.0
loguru>=0.7.3

sqlmodel>=0.0.24
sqlalchemy==2.0.36


langchain==1.2.0
langchain-google-genai==4.0.0
google-generativeai